requirements.txt
sounddevice
scipy
grop
python-dotenv
chromadb
sentence_transformers